# docs/swagger.yaml
definitions:
...
  dto.CreateIdentityServicePayloadDTO:
    properties:
      code:
        type: string
      description:
        type: string
      name:
        type: string
      parent_id:
        type: string
    required:
    - code
    - name
    type: object
...
  dto.IdentityServiceDTO:
    properties:
      id:
        type: string
    type: object
...
  dto.PaginationDTOResponse:
    properties:
      data:
        items: {}
        type: array
      next_page:
        type: integer
      page:
        type: integer
      size:
        type: integer
      total:
        type: integer
    type: object
...
  dto.UpdateIdentityServicePayloadDTO:
    properties:
      code:
        type: string
      description:
        type: string
      name:
        type: string
      parent_id:
        type: string
    required:
    - code
    - name
    type: object
...
  response.ErrorResponse:
    properties:
      code:
        type: string
      errors:
        items: {}
        type: array
      message:
        type: string
      status:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/services:
    get:
      consumes:
      - application/json
      description: Get services
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: size
        type: integer
      - description: Keyword
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successful retrieval of services
          schema:
            $ref: '#/definitions/dto.PaginationDTOResponse'
        "400":
          description: Invalid page number or size
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Retrieve services
      tags:
      - services
    post:
      consumes:
      - application/json
      description: Create a new service
      parameters:
      - description: service payload
        in: body
        name: service
        required: true
        schema:
          $ref: '#/definitions/dto.CreateIdentityServicePayloadDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Successful creation of service
          schema:
            $ref: '#/definitions/dto.IdentityServiceDTO'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Create a new service
      tags:
      - services
  /api/v1/services/{service_id}:
    delete:
      consumes:
      - application/json
      description: Delete an existing service
      parameters:
      - description: Service ID
        in: path
        name: service_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Successful deletion of service
        "400":
          description: Invalid request ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: service not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Delete an existing service
      tags:
      - services
    get:
      consumes:
      - application/json
      description: Get service by ID
      parameters:
      - description: service ID
        in: path
        name: service_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successful retrieval of service
          schema:
            $ref: '#/definitions/dto.IdentityServiceDTO'
        "400":
          description: Invalid request ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: service not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Retrieve service by ID
      tags:
      - services
    put:
      consumes:
      - application/json
      description: Update an existing service
      parameters:
      - description: Service ID
        in: path
        name: service_id
        required: true
        type: string
      - description: service payload
        in: body
        name: service
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateIdentityServicePayloadDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Successful update of service
          schema:
            $ref: '#/definitions/dto.IdentityServiceDTO'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: service not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Update an existing service
      tags:
      - services
...
swagger: "2.0"